Remove LINQPad samples
The LINQPad samples do not compile in LINQPad v5.22.02 when .NET Framework 4.7 is installed (e.g., Windows 10 Creators Update). I am removing the samples until the issue is resolved. The samples are all from the unit test "scenarios" and they work fine in Visual Studio, so just look there (and the docs) for usage/examples.