Fix bug in Stamp.CompareTo
CompareTo must return 0 (same) when two stamps are causally equal or concurrent. The original implementation compared using the Equals method (which, in this implementation is a structural comparison of the entire stamp... the id and event).

This means that 2 stamps that are causally equal will be compared as equal (0) instead of -1 or 1.